Fixed handling of label clicks
Previously, we were triggering clicks on both labels and their associated
inputs when a label was clicked. This had the effect of making it next
to impossible to select multiple choice options in the MC widget.

This patch changes that to only click on labels, not the inputs themselves,
where appropriate. Note that this has the side effect of treating clicks on
inputs enclosed in labels as label clicks effecting the inputs, rather than
input clicks bubbling up to the labels