Fix retarded click supressant not stopping native clicks on the original label
The click supressant for the native "faster-click" worked well mostly,
but because targetElement was being switched out for labels pointing
at inputs, the click supressor was running on the input- but iOS dispatched
the native click on the original label which was clicked.

To fix this, we needed to apply the click supressor to the original click
target, not what we actually dispatch the synthetic click on.

Also, we had to move the capturing handler for this out to the document; for
some reason, binding the capturing handler directly on the element being supressed
caused any bubbling handler on it to run first, which was decidedly suboptimal. Suppressing
the click on the document during capture seems to be suficient to stop this.

Authorized by RM, JP, and BS. Triple the review!